requirements.txt
openai
python-dotenv
pycaw; platform_system == "Windows"
comtypes; platform_system == "Windows"
pyautogui
SpeechRecognition
pyaudio; platform_system == "Windows"
pyttsx3
